# Speed up `totientrange`