This code runs significantly faster in both Python 3.11.3 and PyPy 7.3.12.

```
Python 3.11.3
old: 21.9998 seconds
new: 7.23461 seconds
204% faster

PyPy 7.3.12
old: 11.6819 seconds
new: 5.69909 seconds
105% faster
```

Benchmark code:
```python
import time
import sympy
a = time.time()
list(sympy.sieve.totientrange(10**7, 2 * 10**7))
b = time.time()
print(b - a)
``` 

<!-- BEGIN RELEASE NOTES -->
* ntheory
  * Speed up `totientrange`.
<!-- END RELEASE NOTES -->