@yuanyelele 2023-09-06
The original idea is that $\sum_{d\mid n}\varphi(d)=n$, so $\varphi(n) = n - \sum_{d\mid n, d < n}\varphi(d)$.
The new algorithm is simply based on the fact that $\varphi(n) =n \prod_{p\mid n} \left(1-\frac{1}{p}\right)$.
That is, for each prime number $p$ dividing $n$, let $\varphi(n)$ = $\varphi(n) \cdot (1 - \frac{1}{p})$, or $\varphi(n)$ -= $\varphi(n) / p$.

The speed-up is attributed to the fact that, while division is generally slower than subtraction, only some of the divisors are prime numbers.